Remove CGFloat wrappers around size types. TGMath handles them better

+++ Source/DKSweptAngleGradient.m
@@ -112,7 +112,7 @@
 		cp.y *= 1.5;
 		twopi = 2 * M_PI;
 
-		unsigned long* p = (unsigned long*)buffer;
+		unsigned int* p = (unsigned int*)buffer;
 
 		for (y = 0; y < height; ++y) {
 			for (x = 0; x < width; ++x) {
